add API set user recur food

+++ SEP490_G87_Vita_Nutrient_System_API/SEP490_G87_Vita_Nutrient_System_API/Domain/RequestModels/UserAction.cs
@@ -10,6 +10,6 @@
     {
         public int UserId { get; set; }
         public int FoodId { get; set; }
-        public int RecurId { get; set; }
+        public short RecurId { get; set; }
     }
 }
